+++ src/snippetCommands/pipeSnippet.command.ts
@@ -1,17 +1,19 @@
 import { window, SnippetString } from "vscode";
 import { getFileName, toClassName, toPipeSelector } from "../utils/fileName.parser";
+import { PrefixManager } from "../prefix.manager";
 
 export function insertPipeTemplate() {
   const editor = window.activeTextEditor;
   if (!editor) return;
 
   const fileName = getFileName(editor);
+  const prefix = PrefixManager.getPrefix();
 
 	editor.insertSnippet(new SnippetString(
-`import {  Pipe, PipeTransform } from "@angular/core";
+`import { Pipe, PipeTransform } from "@angular/core";
 
 @Pipe({
-  name: "${toPipeSelector(fileName)}",
+  name: "${toPipeSelector(fileName, prefix)}",
   standalone: true,
 })
 export class ${toClassName(fileName)} implements PipeTransform {